requirements.txt
fastapi
uvicorn
sqlalchemy
databases
asyncpg
psycopg2
psycopg2-binary
alembic
aiomysql
mysqlclient
mysql-connector-python
python-decouple
email-validator
passlib[bcrypt]
pyjwt
python-dotenv
apify-client
regex
boto3
cryptography
python-jose
python-multipart